src/main/webapp/app/entities/rental-my-suffix/delete/rental-my-suffix-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { RentalMySuffixService } from '../service/rental-my-suffix.service';

import { RentalMySuffixDeleteDialogComponent } from './rental-my-suffix-delete-dialog.component';

describe('RentalMySuffix Management Delete Component', () => {
  let comp: RentalMySuffixDeleteDialogComponent;
  let fixture: ComponentFixture<RentalMySuffixDeleteDialogComponent>;
  let service: RentalMySuffixService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [RentalMySuffixDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(RentalMySuffixDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(RentalMySuffixDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(RentalMySuffixService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
